move svg and add them as a svg
