+++ src/pagePartials/home/CertificationProcess.tsx
@@ -1,15 +1,17 @@
-import Image from 'next/image'
-
 import ArrowForwardIosOutlinedIcon from '@mui/icons-material/ArrowForwardIosOutlined'
 import { Box, Typography } from '@mui/material'
+
+import { Choose } from './svg/choose'
+import { Complete } from './svg/complete'
+import { Evidence } from './svg/evidence'
 
 export default function CertificationProcess() {
   return (
     <div>
       <Box display={'flex'} flexDirection={'row'} gap={8} justifyContent={'center'} mb={4}>
-        <Image alt="Choose your certification" height="73" src="shareable/choose.svg" width="116" />
-        <Image alt="Upload the evidence" height="73" src="shareable/evidence.svg" width="116" />
-        <Image alt="Receive your Badge" height="73" src="shareable/complete.svg" width="116" />
+        <Choose />
+        <Evidence />
+        <Complete />
       </Box>
       <Typography
         alignItems={'center'}
