+++ src/pagePartials/home/svg/choose.jsx
