+++ src/pagePartials/home/svg/complete.jsx
